+++ CHANGELOG.md
@@ -2,6 +2,18 @@ This is a list of new fixes and other changes made to SOUP during its continued 
 
 SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd minor versions (1.1.x, 1.3.x, 1.5.x, etc.) are reserved for unstable development versions ("devel" branch on [GitHub][github]), while even ones denote stable releases ("master" branch on [GitHub][github]).  The change log below thus only includes even version numbers, as changes between development versions are too numerous to list.
 
+1.11 (development)
+==================
+
+**New fixes in 1.11.x:**
+
+* [math12803: “Sign up for the newsletter” button overflows the frame on Firefox / Linux](http://meta.math.stackexchange.com/q/12803) (math.SE only)
+
+**Other changes:**
+
+* The fix for [mso215450](http://meta.stackoverflow.com/q/215450) is no longer needed, and has been removed.
+* Fixes can now be more easily limited to or excluded from specific sites.
+
 1.10
 ====